tcctest: add back testXb (self compile with -b)
- Thanks to Kirill "tcc -b itself" should work now
  (was removed in d5f4df09ff4a84dda5b03525285f03be7723376b)

Also:

- tests/Makefile:
  - fix spurious --I from 767410b8750b45d63805b45ca1a2cf34d7cb4923
  - lookup boundtest.c via VPATH (for out-of-tree build)
  - test[123]b?: fail on diff error
  - Windows: test3 now works (from e31579b0769e1f9c0947d12e83316d1149307b1a)
  - abitest: a libtcc.a made by gcc is not usable for tcc
    on WIndows - using source instead (libtcc.c)

- tccpe:
  - avoid gcc warning (x86_64)

+++ tccpe.c
@@ -1803,7 +1803,7 @@ static void pe_add_runtime(TCCState *s1, struct pe_info *pe)
         s1->runtime_main = start_symbol;
 #endif
     } else {
-        pe->start_addr = (DWORD)tcc_get_symbol_err(s1, start_symbol);
+        pe->start_addr = (DWORD)(uintptr_t)tcc_get_symbol_err(s1, start_symbol);
     }
 
     pe->type = pe_type;
